% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data_gateway_by_id.R
\name{access_data_gateway_by_id}
\alias{access_data_gateway_by_id}
\title{Function which retrieves the data from the data access gateway and stores it locally for the shiny app}
\usage{
access_data_gateway_by_id(uid, session, filestore_location,
  file_descriptor = uid, repoID = NULL)
}
\arguments{
\item{uid}{\itemize{
\item the id of the file to retrieve (or NULL if repoid is not NULL)
}}

\item{session}{\itemize{
\item the R Shiny session argument
}}

\item{filestore_location}{\itemize{
\item the path to the filestore
}}

\item{file_descriptor}{\itemize{
\item variable to make any errors more understandable at higher levels, if uid is NULL then this
should be set and the default value not used
}}

\item{repoID}{\itemize{
\item if not NULL then get latest file from this repo id - note exactly one of uid and repoid should be non-NULL
}}
}
\value{
TRUE if successful - throw error if not
}
\description{
This function bypasses the extract modules and allows direct access to the platform. Therefore,
it is your responsibility to ensure errors thrown by this function are caught by the Shiny App.
Setting 'uid' to a valid file uid will bring the file into the filestore and give it the name of the uid, otherwise
setting 'repoID' to a valid repoID will bring the latest file for the given repoID into the file store, and it will be
given the name of the repoID. See the vignette for further details.
}
