% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data_gateway_by_search.R
\name{access_data_gateway_by_search}
\alias{access_data_gateway_by_search}
\title{Function to access the seach API of the BD4BS Shiny sidecar}
\usage{
access_data_gateway_by_search(session, query, tags,
  key_value_query = list(), page = 1, dataset_search = FALSE,
  convert_from_json = !dataset_search, include_obsolete = FALSE)
}
\arguments{
\item{session}{The R shiny session object}

\item{query}{\itemize{
\item the text to search for in the search api call
}}

\item{tags}{\itemize{
\item vector of tags for the call to the search api - set to list() if no tags
}}

\item{key_value_query}{\itemize{
\item list of key values for search (see key_value.R and key_value_filter.R for more details)
}}

\item{page}{\itemize{
\item the API returns 1000 records per page and this argument defines which
page of results we would like to return.
}}

\item{dataset_search}{should the search be for datasets (TRUE) or data files (FALSE - default). If this is set to TRUE then
the include_obsolete argument must be FALSE.}

\item{convert_from_json}{should the function return the json of the response (TRUE) or convert it to R object using
\code{jsonlite::fromJSON} (FALSE)}

\item{include_obsolete}{\itemize{
\item whether to return obsolete files (defaults to FALSE).
In dev-mode the include_obsolete parameter to the search endpoint is ignored
}}
}
\value{
the content of the response (JSON or converted to R objects depending on value of convert_from_json parameter).
The function throws an error if unsuccessful
}
\description{
This function bypasses the extract modules and allows direct access to the platform. Therefore,
it is your responsibility to ensure errors thrown by this function are caught by the Shiny App.
}
