% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data_gateway_dataset_write.R
\name{access_data_gateway_dataset_write}
\alias{access_data_gateway_dataset_write}
\title{Function to write dataset files to BD4BS file store}
\usage{
access_data_gateway_dataset_write(session, name, description = "",
  tags = character(0), access_groups,
  custom_metadata_key_values = list(), datasetID = NULL,
  child_repo_ids = list(), child_dataset_ids = list(),
  child_repo_ids_to_remove = list(),
  child_dataset_ids_to_remove = list())
}
\arguments{
\item{session}{The Shiny session object}

\item{name}{The name of the dataset}

\item{description}{The description of the dataset}

\item{tags}{vector of tags for the dataset}

\item{access_groups}{A vector of AD groups for the dataset.
Note the AD groups of the current user can be retrieved using the "get_ad_groups" function.}

\item{custom_metadata_key_values}{list of custom meta data fields for the dataset or (default) list() if no fields
Each key-value is a list of the form:
Key=unbox("key"), CategoryPath=character(0), Value=unbox("value"), Type=unbox(<<"STRING"|"NUMBER"|"BOOLEAN"|"DATE">>)
where value is a string, number, boolean (true/false) or date (in format as.numeric(as.POSIXct("2014-01-01")) *1000)
see vignette (section about writing files to the platform) for further information}

\item{datasetID}{if NULL (default) then the dataset is created using POST and creates a new dataset, otherwise it updates
a dataset referenced by datasetID.}

\item{child_repo_ids}{a vector of repoIds to be added into the dataset (note NOT dataIDs)}

\item{child_dataset_ids}{a vector of datasetIds to be added into the dataset}

\item{child_repo_ids_to_remove}{a vector of repoIds already in the dataset to be removed (if datasetID is NULL)}

\item{child_dataset_ids_to_remove}{a vector of datasetIds already in the dataset to be removed (if datasetID is NULL)}
}
\value{
The datasetID of the created/updated dataset
}
\description{
Function to write dataset files to BD4BS file store
}
