% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_data_gateway_write.R
\name{access_data_gateway_write}
\alias{access_data_gateway_write}
\title{Function to write file to BD4BS file store}
\usage{
access_data_gateway_write(file, session, name, description = "",
  tags = character(0), access_groups,
  custom_metadata_key_values = list(), file_type = "text/plain",
  repoID = NULL)
}
\arguments{
\item{file}{The path to the file being uploaded to the platform}

\item{session}{The Shiny session object}

\item{name}{The name of the file as referred to in the file store}

\item{description}{The description of the file for the file store}

\item{tags}{Vector of tags for the file}

\item{access_groups}{A vector of AD groups for the file.
Note the AD groups of the current user can be retrieved using the "get_ad_groups" function.}

\item{custom_metadata_key_values}{list of custom meta data fields for the file or (default) list() if no fields
Each key-value is a list of the form:
Key=unbox("key"), CategoryPath=character(0), Value=unbox("value"), Type=unbox(<<"STRING"|"NUMBER"|"BOOLEAN"|"DATE">>)
where value is a string, number, boolean (true/false) or date (in format as.numeric(as.POSIXct("2014-01-01")) *1000)
see vignette for further information}

\item{file_type}{the file type being uploaded (default 'text/plain') if unsure can use the function mime::guess_type()}

\item{repoID}{if NULL (default) then the file is uploaded using POST and creates a new data repository, otherwise it updates
a data entry within the data repository referenced by repoID. See the package vignette for more details}
}
\value{
A named list containing Name, Description, Tags, DataID (the hash of the uploaded file),
RepoID and AccessGroups. If request was not successful then an error is thrown
}
\description{
Function to write file to BD4BS file store
}
