% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_local_filestore.R
\name{access_local_filestore}
\alias{access_local_filestore}
\title{Function which reads file from the local file store}
\usage{
access_local_filestore(filestore_location, file_name,
  file_read_function = readLines, ...)
}
\arguments{
\item{filestore_location}{the path to the filestore}

\item{file_name}{the name of the file}

\item{file_read_function}{the function used to read in the file}

\item{...}{additional arguments to the read function}
}
\value{
the file, or throw an error if cannot read file
}
\description{
Function which reads file from the local file store
}
