% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroy_local_filestore.R
\name{destroy_local_filestore}
\alias{destroy_local_filestore}
\title{Recursively deletes files from given location}
\usage{
destroy_local_filestore(filestore_location)
}
\arguments{
\item{filestore_location}{the path (from the location of the shiny app) to the
filestore location}
}
\value{
0 for success, 1 for failure, invisibly. See unlink function documentation
for further details
}
\description{
Recursively deletes files from given location
}
\details{
This function should be called when shiny session ends to
delete the local file store for the given session.
The following code should be added to server.R:
\code{filestore_location <- setup_local_filestore(session)
onSessionEnded(function(){
 destroy_local_filestore(filestore_location)
})}
}
