% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_platform.R
\name{extract_platform}
\alias{extract_platform}
\title{Server function which corresponds to extract_platformUI}
\usage{
extract_platform(input, output, session, filestore_location)
}
\arguments{
\item{input}{shiny input argument}

\item{output}{shiny output argument}

\item{session}{shiny session argument}

\item{filestore_location}{\itemize{
\item the location to put extracted files
}}
}
\value{
a reactive of the file id of the last extracted file which has been stored in the file store
or NULL if no files extracted/last file attempted to be extracted failed
}
\description{
The function is used to extract
a single user-specified uid from the BD4BS data store and store in the app's filestore
}
