% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_platform_UI.R
\name{extract_platform_UI}
\alias{extract_platform_UI}
\title{Ui module for importing file from BD4BS given file id}
\usage{
extract_platform_UI(id,
  title = "Enter id to extract file from BD4BS platform",
  default_value = "", label_text = "Enter file id")
}
\arguments{
\item{id}{required argument for shiny ui modules}

\item{title}{the title for the modal dialog}

\item{default_value}{what should the id input text box have as a default}

\item{label_text}{what text should be displayed for the input text box}
}
\description{
Ui module for importing file from BD4BS given file id
}
