% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ad_groups.R
\name{get_ad_groups}
\alias{get_ad_groups}
\title{Function which retrieves the AD group names for the user of the app}
\usage{
get_ad_groups(session)
}
\arguments{
\item{session}{\itemize{
\item the R Shiny session argument
}}
}
\value{
vector of user group names for example c("BD4BS-PROD-PUBLIC"). In dev mode this
returns c("dev-group1", "dev-group2")
}
\description{
This function gets the AD group names the current user has access to.
}
