% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_platform_UI.R
\name{search_platform_UI}
\alias{search_platform_UI}
\title{Ui module for searching and importing files from BD4BS}
\usage{
search_platform_UI(id,
  title = "Search BD4BS to import files into this shiny app",
  message = NULL, include_cancel_button = TRUE,
  search_table_style_string = "word-break: break-all; overflow-x:auto; white-space:nowrap;")
}
\arguments{
\item{id}{required argument for shiny ui modules}

\item{title}{the title for the modal dialog}

\item{message}{any message to be displayed above the search input box}

\item{include_cancel_button}{should there be an option to cancel without importing a file?
If TRUE then modalButton is used, If FALSE then no cancel button is displayed and IF "actionButton"
then an action button is displayed and when pressed then the returned value of the module is set to the
time the Cancel button was pressed (i.e. Sys.time()) see corresponding server module for more details}

\item{search_table_style_string}{css character string to handle the display of the search results table
by default this is: \cr 'word-break: break-all; overflow-x:auto; white-space:nowrap;'}
}
\description{
Ui module for searching and importing files from BD4BS
}
