% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_local_filestore.R
\name{setup_local_filestore}
\alias{setup_local_filestore}
\title{Create the filestore for this session's files}
\usage{
setup_local_filestore(session, root_dir = getwd())
}
\arguments{
\item{session}{The R Shiny session object}

\item{root_dir}{The directory in which /tmp/file_store/<<unique session token>>
is to be stored. When running a Shiny App on BD4BS this should be set to
the directory where the server.R/app.R file is located which by default will be the
working directory}
}
\value{
returns path to filestore if successful
throws error if cannot create store
}
\description{
Create the filestore for this session's files
}
\details{
The following code should be added to server.R:
\code{filestore_location <- setup_local_filestore(session)
onSessionEnded(function(){
 destroy_local_filestore(filestore_location)
})}
}
